+++ web/src/plugins/registry.tsx
@@ -3,11 +3,13 @@ import { ClockStatusPlugin } from './ClockStatus';
 import { FlueChatPlugin } from './FlueChat';
 import { GitHubPrListPlugin } from './GitHubPrList';
 import { HostMetricsPlugin } from './HostMetrics';
+import { RuntimeOverviewPlugin } from './RuntimeOverview';
 import type { DisplayPlugin } from '../types';
 
 export const plugins = [
   GitHubPrListPlugin,
   ActiveWatchesPlugin,
+  RuntimeOverviewPlugin,
   FlueChatPlugin,
   HostMetricsPlugin,
   ClockStatusPlugin,